feat: v13 results + annotate 6 new capture chunks from session 20260318
v13 results (46.7 min, 4 experiments):
  - Binary BalAcc=0.709 (RF_bal, 400 feats, class-weighted)
  - Coarse 3-class=0.670 (RF, 400 feats)
  - STATIC/MOTION=0.765 (RF_bal, 400 feats) — STATIC recall 0.573
  - Ceiling-V2 only: weaker (binary=0.664) — early epoch helps
  - Epoch normalization: no improvement (0.665)

New capture session 20260318: 6 chunks x 30s, 30486 CSI packets, 4 nodes
  All chunks annotated with person_count=1, step names for ML labels

+++ scripts/csi_motion_pipeline_v13_ceiling_focused.py
@@ -296,7 +296,7 @@ X_normed = X_all.copy()
 for epoch in df["__epoch"].unique():
     mask = epochs == epoch
     scaler = StandardScaler()
-    X_normed.loc[mask] = scaler.fit_transform(X_normed.loc[mask])
+    X_normed.loc[mask] = scaler.fit_transform(X_normed.loc[mask]).astype(np.float32)
 
 # Use top MI features from global
 for n_feat in [50, 100, 200]:
